Update AddCommand: Shift the code to web section, update to support list of events
