+++ src/main/java/seedu/scheduler/model/ModelManager.java
@@ -13,6 +13,7 @@ import javafx.collections.transformation.FilteredList;
 import seedu.scheduler.commons.core.ComponentManager;
 import seedu.scheduler.commons.core.LogsCenter;
 import seedu.scheduler.commons.events.model.SchedulerChangedEvent;
+import seedu.scheduler.commons.web.ConnectToGoogleCalendar;
 import seedu.scheduler.model.event.Event;
 import seedu.scheduler.model.tag.Tag;
 
@@ -21,9 +22,11 @@ import seedu.scheduler.model.tag.Tag;
  */
 public class ModelManager extends ComponentManager implements Model {
     private static final Logger logger = LogsCenter.getLogger(ModelManager.class);
-
+    private final ConnectToGoogleCalendar connectToGoogleCalendar =
+            new ConnectToGoogleCalendar();
     private final VersionedScheduler versionedScheduler;
     private final FilteredList<Event> filteredEvents;
+
 
     /**
      * Initializes a ModelManager with the given scheduler and userPrefs.
@@ -90,6 +93,7 @@ public class ModelManager extends ComponentManager implements Model {
     @Override
     public void addEvents(List<Event> events) {
         versionedScheduler.addEvents(events);
+        connectToGoogleCalendar.pushToGoogleCal(events);
         updateFilteredEventList(PREDICATE_SHOW_ALL_EVENTS);
         indicateSchedulerChanged();
     }
